update jwt filter to prevent query db, add retry and sleep for 5s to prevent api exhaustion from Hugging Face.

+++ backend/src/main/java/com/example/bbmovie/model/elasticsearch/MovieDocument.java
@@ -21,8 +21,8 @@ public class MovieDocument {
     @Field(type = FieldType.Text, analyzer = "standard")
     private String description;
 
-    @Field(type = FieldType.Double)
-    private Double rating;
+    @Field(type = FieldType.Float)
+    private Float rating;
 
     @Field(type = FieldType.Keyword)
     private List<String> categories;
